faet(index): change main index to ts
